# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Order creator
  description: Order creator
  version: 1.0.0
tags:
  - name: orders
paths:
  /orders:
    post:
      tags:
        - orders
      description: Creates an order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/order-request-dto'
      responses:
        '201':
          description: The order has been created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order-response-dto'
    get:
      tags:
        - orders
      description: Retrieve list of orders
      operationId: getAllOrders
      parameters:
        - in: query
          name: pageable
          schema:
            $ref: '#/components/schemas/pageable'
          required: true
          description: "pageable"
      responses:
        '200':
          description: The order has been created successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/order-response-dto'
  /orders/{orderId}:
    get:
      tags:
        - orders
      description: Get an order by ID.
      operationId: getOrderById
      parameters:
        - in: path
          name: orderId
          schema:
            type: string
          required: true
          description: "orderId"
          example: "00000000-0000-0000-c000-000000000001"
      responses:
        '200':
          description: Return order by a given ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order-response-dto'
    delete:
      tags:
        - orders
      description: Delete an order by ID.
      operationId: deleteOrderById
      parameters:
        - in: path
          name: orderId
          schema:
            type: string
          required: true
          description: "orderId"
          example: "00000000-0000-0000-c000-000000000001"
      responses:
        '204':
          description: Delete order by a given ID.

    put:
      tags:
        - orders
      description: Update an order by ID.
      operationId: updateOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/order-update-dto'
      parameters:
        - in: path
          name: orderId
          schema:
            type: string
          required: true
          description: "orderId"
          example: "00000000-0000-0000-c000-000000000001"
      responses:
        '200':
          description: Update order by a given ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order-response-dto'
components:
  schemas:
    order-request-dto:
      title: OrderRequestDto
      required:
        - orderName
        - quantity
        - price
        - priority
      type: object
      properties:
        orderName:
          type: string
          minLength: 1
          maxLength: 50
          example: new Order
        quantity:
          type: integer
          minLength: 1
          maxLength: 100
          example: 1
        price:
          type: integer
          minLength: 1
          example: 100
        priority:
          type: integer
          minLength: 1
          maxLength: 5
          example: 1

    order-response-dto:
      title: OrderResponseDto
      required:
        - orderName
        - orderId
        - quantity
        - price
        - priority
      type: object
      properties:
        orderName:
          type: string
          minLength: 1
          maxLength: 50
          example: new Order
        orderId:
          type: string
          minLength: 1
          maxLength: 36
          example: 00000000-0000-0000-c000-000000000001
        quantity:
          type: integer
          minLength: 1
          maxLength: 100
          example: 1
        price:
          type: integer
          minLength: 1
          example: 100
        priority:
          type: integer
          minLength: 1
          maxLength: 5
          example: 1

    order-update-dto:
      title: OrderUpdateDto
      required:
        - orderName
        - quantity
        - price
        - priority
      type: object
      properties:
        orderName:
          type: string
          minLength: 1
          maxLength: 50
          example: new Order
        quantity:
          type: integer
          minLength: 1
          maxLength: 100
          example: 1
        price:
          type: integer
          minLength: 1
          example: 100
        priority:
          type: integer
          minLength: 1
          maxLength: 5
          example: 1
    pageable:
      title: pageable
      required:
        - from
        - size
        - sort
      type: object
      properties:
        from:
          type: integer
          minLength: 0
          maxLength: 50
          example: 1
        size:
          type: integer
          minLength: 10
          maxLength: 100
          example: 1
        sort:
          type: string
          minLength: 1
          example: ASC